requirements.txt
openai-whisper
pyaudio
numpy
